+++ src/libs/TransactionRecoveryApi/constants/chain/mainnet.ts
@@ -9,6 +9,7 @@ export const rpcMap: Record<EvmChain, string> = {
   [EvmChain.ETHEREUM]: "https://mainnet.infura.io/v3/510b6d5b3c56497b8070626a54f565a9",
   [EvmChain.AURORA]: "https://mainnet.aurora.dev",
   [EvmChain.BINANCE]: "https://bsc-dataseed.binance.org",
+  [EvmChain.BNBCHAIN]: "https://bsc-dataseed.binance.org",
   [EvmChain.ARBITRUM]: "https://arb1.arbitrum.io/rpc",
   [EvmChain.CELO]: "https://forno.celo.org",
   [EvmChain.KAVA]: "https://evm.kava.io",
@@ -42,6 +43,10 @@ export const networkInfo: Record<EvmChain, Network> = {
   [EvmChain.BINANCE]: {
     chainId: 56,
     name: EvmChain.BINANCE,
+  },
+  [EvmChain.BNBCHAIN]: {
+    chainId: 56,
+    name: EvmChain.BNBCHAIN,
   },
   [EvmChain.ARBITRUM]: {
     chainId: 42161,