+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - adding feature to generate deposit addresses "offline"
 - AxelarGMPRecoveryAPI
   - fixes to manualRelayToDestinationChain to first check if transaction is already confirmed but not broadcasted, and broadcast the transaction (as identified by command ID) if so
+- update constants to include `BNBCHAIN` (in addition to `BINANCE`)
 
 ## [0.12.4] - 2023-FEBRUARY-1